results-compilation/requirements.txt
pandas<2.0
fastplot
pyyaml